refactor: Use new panelheader accros all components
