+++ projects/hslayers/src/components/draw/draw.module.ts
@@ -9,12 +9,13 @@ import {DrawPanelComponent} from './draw-panel/draw-panel.component';
 import {HsDrawComponent} from './draw.component';
 import {HsDrawLayerMetadataDialogComponent} from './draw-layer-metadata/draw-layer-metadata.component';
 import {HsDrawToolbarComponent} from './draw-toolbar/draw-toolbar.component';
-import {TranslateCustomPipe} from '../language/translate-custom.pipe';
 import {HsLaymanModule} from '../../common/layman/layman.module';
+import {HsPanelHeaderComponent} from '../layout/panels/panel-header/panel-header.component';
 import {HsPanelHelpersModule} from '../layout/panels/panel-helpers.module';
 import {HsQueryModule} from '../query/query.module';
 import {HsRmLayerDialogModule} from '../../common/remove-multiple/remove-layer-dialog.module';
 import {HsStylerModule} from '../styles/styles.module';
+import {TranslateCustomPipe} from '../language/translate-custom.pipe';
 
 @NgModule({
   schemas: [CUSTOM_ELEMENTS_SCHEMA],
@@ -35,6 +36,7 @@ import {HsStylerModule} from '../styles/styles.module';
     HsQueryModule,
     HsLaymanModule,
     HsRmLayerDialogModule,
+    HsPanelHeaderComponent,
   ],
   exports: [
     HsDrawComponent,
