+++ projects/hslayers/src/components/feature-table/feature-table.module.ts
@@ -3,10 +3,11 @@ import {CommonModule} from '@angular/common';
 import {FormsModule} from '@angular/forms';
 import {HsFeatureFilterPipe} from './feature-filter.pipe';
 import {HsFeatureTableComponent} from './feature-table.component';
-import {TranslateCustomPipe} from '../language/translate-custom.pipe';
 import {HsLayerFeaturesComponent} from './layer-features.component';
+import {HsPanelHeaderComponent} from '../layout/panels/panel-header/panel-header.component';
 import {HsPanelHelpersModule} from '../layout/panels/panel-helpers.module';
 import {NgbDropdownModule} from '@ng-bootstrap/ng-bootstrap';
+import {TranslateCustomPipe} from '../language/translate-custom.pipe';
 
 @NgModule({
   schemas: [CUSTOM_ELEMENTS_SCHEMA],
@@ -21,6 +22,7 @@ import {NgbDropdownModule} from '@ng-bootstrap/ng-bootstrap';
     HsPanelHelpersModule,
     TranslateCustomPipe,
     NgbDropdownModule,
+    HsPanelHeaderComponent,
   ],
   exports: [
     HsFeatureTableComponent,
