+++ projects/hslayers/src/components/language/language.component.html
@@ -1,5 +1,5 @@
 <div class="card hs-main-panel" [hidden]="(isVisible$ | async) === false">
-    <hs-panel-header name="language" [title]="'PANEL_HEADER.LANGUAGE' | translateHs">
+    <hs-panel-header name="language" [panelTabs]="'LANGUAGE'">
     </hs-panel-header>
     <div class="card-body">
         <ul class="m-3 list-group m-auto w-75 py-md-2">
