+++ projects/hslayers/src/components/legend/legend.component.html
@@ -1,5 +1,5 @@
 <div class="card hs-main-panel" [hidden]="(isVisible$ | async) === false">
-    <hs-panel-header name="legend" [title]="'PANEL_HEADER.LEGEND' | translateHs">
+    <hs-panel-header name="legend" [panelTabs]="'LEGEND'">
     </hs-panel-header>
     <div class="card-body">
         <p><input type="text" class="form-control hs-filter" [placeholder]="'COMMON.filter' | translateHs"
