+++ projects/hslayers/src/components/measure/measure.module.ts
@@ -3,10 +3,11 @@ import {CommonModule} from '@angular/common';
 import {FormsModule} from '@angular/forms';
 import {NgbDropdownModule} from '@ng-bootstrap/ng-bootstrap';
 
-import {TranslateCustomPipe} from '../language/translate-custom.pipe';
 import {HsMeasureComponent} from './measure.component';
 import {HsMeasureToolbarComponent} from './measure-toolbar.component';
+import {HsPanelHeaderComponent} from '../layout//panels/panel-header/panel-header.component';
 import {HsPanelHelpersModule} from '../layout/panels/panel-helpers.module';
+import {TranslateCustomPipe} from '../language/translate-custom.pipe';
 
 @NgModule({
   schemas: [CUSTOM_ELEMENTS_SCHEMA],
@@ -17,6 +18,7 @@ import {HsPanelHelpersModule} from '../layout/panels/panel-helpers.module';
     HsPanelHelpersModule,
     NgbDropdownModule,
     TranslateCustomPipe,
+    HsPanelHeaderComponent,
   ],
   exports: [HsMeasureComponent, HsMeasureToolbarComponent],
 })
