+++ projects/hslayers/src/components/permalink/share.component.html
@@ -1,5 +1,5 @@
 <div class="card panel-default hs-main-panel" [hidden]="(isVisible$ | async) === false">
-    <hs-panel-header name="permalink" [title]="'PANEL_HEADER.PERMALINK' | translateHs">
+    <hs-panel-header name="permalink" [panelTabs]="'PERMALINK'">
     </hs-panel-header>
     <div class="card-body">
         <div class="card-body">
