+++ projects/hslayers/src/components/permalink/share.module.ts
@@ -7,14 +7,21 @@ import {
 import {CUSTOM_ELEMENTS_SCHEMA, NgModule} from '@angular/core';
 import {FormsModule} from '@angular/forms';
 
-import {TranslateCustomPipe} from '../language/translate-custom.pipe';
+import {HsPanelHeaderComponent} from '../layout//panels/panel-header/panel-header.component';
 import {HsPanelHelpersModule} from '../layout/panels/panel-helpers.module';
 import {HsShareComponent} from './share.component';
+import {TranslateCustomPipe} from '../language/translate-custom.pipe';
 
 @NgModule({
   schemas: [CUSTOM_ELEMENTS_SCHEMA],
   declarations: [HsShareComponent],
-  imports: [CommonModule, FormsModule, HsPanelHelpersModule, TranslateCustomPipe],
+  imports: [
+    CommonModule,
+    FormsModule,
+    HsPanelHelpersModule,
+    TranslateCustomPipe,
+    HsPanelHeaderComponent,
+  ],
   exports: [HsShareComponent],
   providers: [
     {provide: LocationStrategy, useClass: PathLocationStrategy},
