+++ projects/hslayers/src/components/print/print.module.ts
@@ -5,13 +5,14 @@ import {FormsModule} from '@angular/forms';
 import {ColorSketchModule} from 'ngx-color/sketch';
 import {NgbDropdownModule} from '@ng-bootstrap/ng-bootstrap';
 
-import {TranslateCustomPipe} from '../language/translate-custom.pipe';
+import {HsPanelHeaderComponent} from '../layout/panels/panel-header/panel-header.component';
 import {HsPanelHelpersModule} from '../layout/panels/panel-helpers.module';
 import {HsPrintComponent} from './print.component';
 import {HsPrintImprintStylerComponent} from './imprint-styler/imprint-styler.component';
 import {HsPrintLegendStylerComponent} from './legend-styler/legend-styler.component';
 import {HsPrintScaleStylerComponent} from './scale-styler/scale-styler.component';
 import {HsPrintTextStylerComponent} from './text-styler/text-styler.component';
+import {TranslateCustomPipe} from '../language/translate-custom.pipe';
 
 /**
  * Add print dialog template to the app
@@ -32,6 +33,7 @@ import {HsPrintTextStylerComponent} from './text-styler/text-styler.component';
     TranslateCustomPipe,
     ColorSketchModule,
     NgbDropdownModule,
+    HsPanelHeaderComponent,
   ],
   exports: [
     HsPrintComponent,
