+++ projects/hslayers/src/components/save-map/save-map.module.ts
@@ -3,14 +3,15 @@ import {CommonModule} from '@angular/common';
 import {FormsModule, ReactiveFormsModule} from '@angular/forms';
 
 import {AdvancedOptionsComponent} from './form/advanced-options/advanced-options.component';
-import {TranslateCustomPipe} from '../language/translate-custom.pipe';
 import {HsLaymanModule} from '../../common/layman/layman.module';
 import {HsLogModule} from '../../common/log/log.module';
+import {HsPanelHeaderComponent} from '../layout/panels/panel-header/panel-header.component';
 import {HsPanelHelpersModule} from '../layout/panels/panel-helpers.module';
 import {HsSaveMapAdvancedFormComponent} from './form/form.component';
 import {HsSaveMapComponent} from './save-map.component';
 import {HsSaveMapResultDialogComponent} from './dialog-result/dialog-result.component';
 import {HsUiExtensionsModule} from '../../common/widgets/ui-extensions.module';
+import {TranslateCustomPipe} from '../language/translate-custom.pipe';
 
 @NgModule({
   schemas: [CUSTOM_ELEMENTS_SCHEMA],
@@ -29,6 +30,7 @@ import {HsUiExtensionsModule} from '../../common/widgets/ui-extensions.module';
     HsUiExtensionsModule,
     TranslateCustomPipe,
     HsLaymanModule,
+    HsPanelHeaderComponent,
   ],
   exports: [HsSaveMapComponent],
 })
