+++ projects/hslayers/src/components/search/search.module.ts
@@ -2,12 +2,13 @@ import {CUSTOM_ELEMENTS_SCHEMA, NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
 import {FormsModule} from '@angular/forms';
 
-import {TranslateCustomPipe} from '../language/translate-custom.pipe';
+import {HsPanelHeaderComponent} from '../layout/public-api';
 import {HsPanelHelpersModule} from '../layout/panels/panel-helpers.module';
 import {HsSearchComponent} from './search.component';
 import {HsSearchInputComponent} from './search-input.component';
 import {HsSearchResultsComponent} from './search-results.component';
 import {HsSearchToolbarComponent} from './search-toolbar.component';
+import {TranslateCustomPipe} from '../language/translate-custom.pipe';
 import {limitToPipe} from './limitTo.pipe';
 
 @NgModule({
@@ -19,7 +20,13 @@ import {limitToPipe} from './limitTo.pipe';
     limitToPipe,
     HsSearchToolbarComponent,
   ],
-  imports: [FormsModule, CommonModule, HsPanelHelpersModule, TranslateCustomPipe],
+  imports: [
+    FormsModule,
+    CommonModule,
+    HsPanelHelpersModule,
+    TranslateCustomPipe,
+    HsPanelHeaderComponent,
+  ],
   exports: [
     HsSearchComponent,
     HsSearchResultsComponent,
