+++ projects/hslayers/src/css/hslayers-bootstrap.scss
@@ -747,8 +747,8 @@ $list-group-item-padding-x: 1.25rem;
     margin-right: 0.5em;
   }
 
-  .hs-extra-button {
-    margin-right: 0.5em;
+  extra-buttons a.dropdown-item i {
+    padding-right: 0.2rem;
   }
 
   .dropdown-toggle.hs-custom-toggle::after {
