+++ projects/hslayers/src/css/hslayers-wo-bootstrap.scss
@@ -605,10 +605,6 @@ $list-group-item-padding-x: 1.25rem;
     margin-right: 0.5em;
   }
 
-  .hs-extra-button {
-    margin-right: 0.5em;
-  }
-
   .dropdown-toggle.hs-custom-toggle::after {
     display: inline-block;
     width: 0.47em;
